requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-filter==21.1
celery[redis]==5.2.7
flower==1.2.0
